Implement - Carousel Items form
I have to create a new carousel service in order to maintain modularity.

Now carousel is a standalone component.

added to carousel an @input in order to enable or disable the new navigate linked to the images.

Still in work , i need to implement a few of validations and test it properly